famoir-backend/requirements.txt
google-adk
google-genai
fastapi
uvicorn[standard]
websockets
python-dotenv
pydantic
firebase-admin
reportlab
